Extract Discord firehose chat behavior to a hook
Pluggable actions to take with optional integrations when a chat message is
created seems like exactly the sort of use case the hook system was designed
to support.

Shifting the code there made it easy to notice and fix a couple of bugs:

* The firehose did not respect the `disable.discord` flag.
* System ChatMessages created in response to guesses transitioning states were
  double-created.
