+++ imports/server/hooks/hookset.ts
@@ -20,6 +20,10 @@ interface Hookset {
   // Triggered when a puzzle that was marked solved is marked unsolved (by a
   // guess previously marked correct being unwound to a different state).
   onPuzzleNoLongerSolved?: (puzzleId: string) => void;
+
+  // Triggered when a new message is added to a puzzle's chat (either from a
+  // user or e.g. in response to a guess transitioning state).
+  onChatMessageCreated?: (chatMessageId: string) => void;
 }
 
 export default Hookset;
